Migrate work database service into new work subdirectory of the app directory - Update file for functional preference over class based
